removed memoized selector selectCartItems and store whole product in cart state
Removed it because the memoized selector no longer outputs correct.
selectCartItems is dependent on fetchedProducts state but since the products changes everytime we search for a new product the selectCartItems becomes empty.

The new solution is to simply store the whole object of the product to "remember" it in cart state instead of storing only product ids.